Create CardRemove.bat
This file is needed for Windows users of the Yubikey when working with Gpg4win. When removing the Yubikey, the Windows Smart Card service seems to cause the gpg-agent to no longer talk to the Yubikey. Removing the Yubikey, stopping the gpg-agent, stopping the Smart Card service and starting the gpg-agent again seems to reliably restore functionality when the Yubikey is plugged back in.